+++ internal/validate/workloads.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"github.com/hrodrig/kzero/internal/config"
-	"github.com/hrodrig/kzero/internal/log"
 	"github.com/hrodrig/kzero/internal/executor"
+	"github.com/hrodrig/kzero/internal/log"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"